Re-integrate cozy changes
I broke everything by merging from master, which contained a commit not
in a released version.
So I had to revert the cozy branch to a previous state, but that's
advanced giting and I'm not very good at this so I kind of wiped out the
whole thing. Anyway, should be good again.